Add 4 to RPi firmware files to avoid confusion
The Raspberry Pi 4 start.elf and fixup.dat have 4's in them with
Raspberry Pi OS.

The `config.txt` says what to use so the current setup works.

However, there's a sentence in the
[docs](https://www.raspberrypi.com/documentation/computers/config_txt.html#start_x-start_debug)
that says "On the Raspberry Pi 4, if the files start4x.elf and
fixup4x.dat are present, these files will be used instead [of
start_x.elf and fixup_x.dat)."

That means that if you copy `start4.elf` to the FAT filesystem and leave
the provided `start.elf` there, that `start4.elf` is used in preference
to one in the `config.txt`. This makes sense, but is cause confusion.

Therefore, to avoid any confusion or surprises, add the 4's back to the
names. It's still not 1:1 with the RaspberryPi OS given that the x or
non-x firmware is decided by which file is sourced by the `fwup.conf`,
but it's consistent with other Nerves RPi systems and doesn't have 4
confusion.